Rename solvers/trajectoryOptimization to solvers/trajectory_optimization
