+++ drake/solvers/CMakeLists.txt
@@ -131,4 +131,4 @@ if(BUILD_TESTING)
   add_subdirectory(test)
 endif()
 
-add_subdirectory(trajectoryOptimization)
+add_subdirectory(trajectory_optimization)
