+++ drake/solvers/trajectory_optimization/CMakeLists.txt
