+++ drake/solvers/trajectory_optimization/direct_trajectory_optimization.h
