+++ drake/solvers/trajectory_optimization/test/CMakeLists.txt
