+++ drake/solvers/trajectory_optimization/test/trajectory_optimization_test.cc
@@ -4,7 +4,7 @@
 
 #include "drake/common/eigen_matrix_compare.h"
 #include "drake/common/trajectories/piecewise_polynomial.h"
-#include "drake/solvers/trajectoryOptimization/direct_trajectory_optimization.h"
+#include "drake/solvers/trajectory_optimization/direct_trajectory_optimization.h"
 
 using std::vector;
 using Eigen::MatrixXd;
